requirements/base.txt
# Django
Django==1.9

# Environmental var configuration
django-configurations==1.0

# For the persistence stores
psycopg2==2.6.1
dj-database-url==0.3.0

# Models
django-model-utils==2.4

# Rest apis
djangorestframework==3.3.2
djangorestframework-gis==0.10.1
Markdown==2.6.5
django-filter==0.11.0
djoser==0.4.2

# Static files
whitenoise==2.0.6

# Media files
boto==2.38.0
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.2

# For asynchronous queuing
django-rq==0.9.0

# Time zones support
pytz==2015.7

# Used for currencies
py-moneyed==0.6.0
django-money==0.7.6